@@ -21,6 +21,7 @@
 		"@playwright/test": "^1.49.1",
 		"@plone/blocks": "1.0.0-alpha.3",
 		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/adapter-node": "^5.4.0",
 		"@sveltejs/kit": "^2.16.0",
 		"@sveltejs/vite-plugin-svelte": "^5.0.0",
 		"@tailwindcss/forms": "^0.5.9",
@@ -31,7 +32,7 @@
 		"jsdom": "^26.0.0",
 		"prettier": "^3.4.2",
 		"prettier-plugin-svelte": "^3.3.3",
-		"svelte": "^5.0.0",
+		"svelte": "^5.46.1",
 		"svelte-check": "^4.0.0",
 		"svelte-preprocess-react": "^2.1.0",
 		"tailwindcss": "^4.0.0",
