@@ -3,6 +3,13 @@ import '@testing-library/jest-dom/vitest';
 import { render, screen } from '@testing-library/svelte';
 import SlateBlockView from './SlateBlockView.svelte';
 
+const defaultProps = {
+	metadata: {},
+	properties: {},
+	path: '/',
+	blocksConfig: {}
+};
+
 // Task Group 1: Paragraph and default element handling tests
 describe('Slate Block - Paragraph Element', () => {
 	test('renders paragraph with correct p element', () => {
@@ -16,7 +23,7 @@ describe('Slate Block - Paragraph Element', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const paragraph = container.querySelector('p');
@@ -35,7 +42,7 @@ describe('Slate Block - Paragraph Element', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const paragraph = container.querySelector('p.slate-paragraph');
@@ -53,7 +60,7 @@ describe('Slate Block - Paragraph Element', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		// Should render as paragraph (fallback)
@@ -80,7 +87,7 @@ describe('Slate Block - Paragraph Element', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const paragraph = container.querySelector('p.slate-paragraph');
@@ -106,7 +113,7 @@ describe('Slate Block - Heading Elements', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const heading = container.querySelector('h4.slate-h4');
@@ -125,7 +132,7 @@ describe('Slate Block - Heading Elements', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const heading = container.querySelector('h5.slate-h5');
@@ -144,7 +151,7 @@ describe('Slate Block - Heading Elements', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const heading = container.querySelector('h6.slate-h6');
@@ -167,7 +174,7 @@ describe('Slate Block - Heading Elements', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		expect(container.querySelector('h2.slate-h2')).toBeInTheDocument();
@@ -193,7 +200,7 @@ describe('Slate Block - Inline Text Formatting', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const underline = container.querySelector('u.slate-underline');
@@ -217,7 +224,7 @@ describe('Slate Block - Inline Text Formatting', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const strikethrough = container.querySelector('s.slate-strikethrough');
@@ -241,7 +248,7 @@ describe('Slate Block - Inline Text Formatting', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const del = container.querySelector('del.slate-del');
@@ -272,7 +279,7 @@ describe('Slate Block - Inline Text Formatting', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const del = container.querySelector('del.slate-del');
@@ -302,7 +309,7 @@ describe('Slate Block - Inline Text Formatting', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const sub = container.querySelector('sub.slate-sub');
@@ -327,7 +334,7 @@ describe('Slate Block - Inline Text Formatting', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const sup = container.querySelector('sup.slate-sup');
@@ -353,7 +360,7 @@ describe('Slate Block - Inline Text Formatting', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const code = container.querySelector('code.slate-code');
@@ -381,7 +388,7 @@ describe('Slate Block - Inline Text Formatting', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		expect(container.querySelector('strong.slate-strong')).toBeInTheDocument();
@@ -407,7 +414,7 @@ describe('Slate Block - Blockquote and Links', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const blockquote = container.querySelector('blockquote.slate-blockquote');
@@ -432,7 +439,7 @@ describe('Slate Block - Blockquote and Links', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const link = container.querySelector('a.slate-link');
@@ -458,7 +465,7 @@ describe('Slate Block - Blockquote and Links', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const link = container.querySelector('a.slate-link');
@@ -495,7 +502,7 @@ describe('Slate Block - Blockquote and Links', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const link = container.querySelector('a.slate-link');
@@ -527,7 +534,7 @@ describe('Slate Block - Lists', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const ol = container.querySelector('ol.slate-ol');
@@ -553,7 +560,7 @@ describe('Slate Block - Lists', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const ul = container.querySelector('ul.slate-ul');
@@ -587,7 +594,7 @@ describe('Slate Block - Lists', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const outerUl = container.querySelector('ul.slate-ul');
@@ -633,7 +640,7 @@ describe('Slate Block - Lists', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		expect(container.querySelector('h2.slate-h2')).toBeInTheDocument();
@@ -669,7 +676,7 @@ describe('Slate Block - Integration & Edge Cases', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const em = container.querySelector('em.slate-em');
@@ -707,7 +714,7 @@ describe('Slate Block - Integration & Edge Cases', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const blockquote = container.querySelector('blockquote.slate-blockquote');
@@ -743,7 +750,7 @@ describe('Slate Block - Integration & Edge Cases', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const ol = container.querySelector('ol.slate-ol');
@@ -759,7 +766,7 @@ describe('Slate Block - Integration & Edge Cases', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const slateBlock = container.querySelector('.block.slate');
@@ -789,7 +796,7 @@ describe('Slate Block - Integration & Edge Cases', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const link = container.querySelector('a.slate-link');
@@ -822,7 +829,7 @@ describe('Slate Block - Integration & Edge Cases', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const li = container.querySelector('li.slate-li');
@@ -850,7 +857,7 @@ describe('Slate Block - Integration & Edge Cases', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const h3 = container.querySelector('h3.slate-h3');
@@ -897,7 +904,7 @@ describe('Slate Block - Integration & Edge Cases', () => {
 		};
 
 		const { container } = render(SlateBlockView, {
-			props: { key: 'test-key', id: 'test-id', data }
+			props: { ...defaultProps, key: 'test-key', id: 'test-id', data }
 		});
 
 		const blockquote = container.querySelector('blockquote.slate-blockquote');
