@@ -5,7 +5,21 @@ import Page from './+page.svelte';
 
 describe('/+page.svelte', () => {
 	test('should render h1', () => {
-		render(Page);
+		const mockData = {
+			title: 'Test Title',
+			'@id': 'http://localhost:8080/Plone',
+			blocks: {},
+			blocks_layout: { items: [] },
+			listingData: {},
+			listingPages: {},
+			paginatedBlockCount: 0
+		};
+
+		render(Page, {
+			props: {
+				data: mockData
+			}
+		});
 		expect(screen.getByRole('heading', { level: 1 })).toBeInTheDocument();
 	});
 });
